// src/app/core/models/interfaces/ICustommerDetailes.ts
export interface ICustomerDetailes {
    customerID: number;
    userid :any;
    custCardNo: string;
    custommerImage: string;
    custommerSignature: string;
    companyId: number;
    fullName: string;
    gender: string;
    dateOfBirth: string;
    nationlityId: number;
    nationality: string;
    maritalStatus: string;
    occupation: string;
    educationLevel: string | null;
    nationalIDOrPassport: string;
    taxIdentificationNumber: string;
    drivingLicenseNumber: string;
    phoneNumber?: string | null;
    alternativePhoneNumber?: string | null;
    emailAddress?: string | null;
    preStreet?: string | null;
    perStreet?: string | null;
    preZIP?: string | null;
    perZIP?: string | null;
    preCity?: string | null;
    perCity?: string | null;
    preState?: string | null;
    perState?: string | null;
    employmentType?: string | null;
    employerOrBusnName?: string | null;
    jobTitleOrBusnType?: string | null;
    monthlyIncOrBusnRev: number;
    yearsOfExpOrBusnAge: number;
    workOrBusnAddress?: string | null;
    employerOrBusnContact?: string | null;
    bankName?: string | null;
    accountNumber?: string | null;
    monthlyIncomeSources: number;
    monthlyExpenses: number;
    assetsOwned?: string | null;
    liabilities?: string | null;
    guarantorImage?: string | null;
    guarantorFullName?: string | null;
    relationshipWithApplicant?: string | null;
    guarantorContactNumber?: string | null;
    guarantorAddress?: string | null;
    guarantorNationalIDOrPassport?: string | null;
    guarantorSignature?: string | null;
  }
  